started extracting global functions and effects
