+++ src/context/ContextProviders.tsx
@@ -1,4 +1,4 @@
-import { createContext, PropsWithChildren, useContext, useRef, useState } from 'react';
+import { createContext, PropsWithChildren, useContext, useMemo, useRef, useState } from 'react';
 import { ContextMenuProvider } from './ContextMenuContext';
 import { ClerkProvider } from '@clerk/clerk-react';
 import { CLERK_PUBLISHABLE_KEY } from '../utils/constants';
@@ -255,6 +255,24 @@ function createAppStateContext() {
   const dependants = {
     ...useContextState<Panel | null>()('renderedRightPanel', state.activeRightPanel),
     currentFolderPathRef: useRef<string>(state.currentFolderPath),
+    isLightTheme: useMemo(() => [Theme.Light, Theme.Snow, Theme.Arctic].includes(state.theme as Theme), [state.theme]),
+    geometricAdjustmentsKey: useMemo(() => {
+      if (!state.adjustments) return '';
+      const { crop, rotation, flipHorizontal, flipVertical, orientationSteps } = state.adjustments;
+      return JSON.stringify({ crop, rotation, flipHorizontal, flipVertical, orientationSteps });
+    }, [
+      state.adjustments?.crop,
+      state.adjustments?.rotation,
+      state.adjustments?.flipHorizontal,
+      state.adjustments?.flipVertical,
+      state.adjustments?.orientationSteps,
+    ]),
+    visualAdjustmentsKey: useMemo(() => {
+      if (!state.adjustments) return '';
+      const { rating: _rating, sectionVisibility: _sectionVisibility, ...visualAdjustments } = state.adjustments;
+      return JSON.stringify(visualAdjustments);
+    }, [state.adjustments]),
+    pinnedFolders: useMemo(() => state.appSettings?.pinnedFolders || [], [state.appSettings]),
   };
 
   return {
