+++ src/hooks/handlers/useCreateResizeHandler.ts
@@ -1,0 +1,32 @@
+import { useAppState } from '../../context/ContextProviders';
+
+export function useCreateResizeHandler() {
+  const { setIsResizing, setLeftPanelWidth, setRightPanelWidth, setBottomPanelHeight } = useAppState();
+
+  const createResizeHandler = (setter: any, startSize: number) => (e: any) => {
+    e.preventDefault();
+    setIsResizing(true);
+    const startX = e.clientX;
+    const startY = e.clientY;
+    const doDrag = (moveEvent: any) => {
+      if (setter === setLeftPanelWidth) {
+        setter(Math.max(200, Math.min(startSize + (moveEvent.clientX - startX), 500)));
+      } else if (setter === setRightPanelWidth) {
+        setter(Math.max(280, Math.min(startSize - (moveEvent.clientX - startX), 600)));
+      } else if (setter === setBottomPanelHeight) {
+        setter(Math.max(100, Math.min(startSize - (moveEvent.clientY - startY), 400)));
+      }
+    };
+    const stopDrag = () => {
+      document.documentElement.style.cursor = '';
+      window.removeEventListener('mousemove', doDrag);
+      window.removeEventListener('mouseup', stopDrag);
+      setIsResizing(false);
+    };
+    document.documentElement.style.cursor = setter === setBottomPanelHeight ? 'row-resize' : 'col-resize';
+    window.addEventListener('mousemove', doDrag);
+    window.addEventListener('mouseup', stopDrag);
+  };
+
+  return createResizeHandler;
+}
