+++ src/hooks/handlers/useDebounce.ts
@@ -1,0 +1,20 @@
+import { invoke } from '@tauri-apps/api/core';
+import { useCallback } from 'react';
+import { Invokes } from '../../components/ui/AppProperties';
+import { useAppState } from '../../context/ContextProviders';
+import debounce from 'lodash.debounce';
+
+export function useDebouncedSave() {
+  const { setError } = useAppState();
+  const debouncedSave = useCallback(
+    debounce((path, adjustmentsToSave) => {
+      invoke(Invokes.SaveMetadataAndUpdateThumbnail, { path, adjustments: adjustmentsToSave }).catch((err) => {
+        console.error('Auto-save failed:', err);
+        setError(`Failed to save changes: ${err}`);
+      });
+    }, 300),
+    [],
+  );
+
+  return debouncedSave;
+}
