+++ src/hooks/handlers/useHandleGenerativeReplace.ts
@@ -1,0 +1,89 @@
+import { invoke } from '@tauri-apps/api/core';
+import { useCallback } from 'react';
+import { Invokes } from '../../components/ui/AppProperties';
+import { AiPatch, Adjustments } from '../../utils/adjustments';
+import { useAppState } from '../../context/ContextProviders';
+import { useSetAdjustments } from './useSetAdjustments';
+
+export function useHandleGenerativeReplace() {
+  const {
+    selectedImage,
+    isGeneratingAi,
+    adjustments,
+    setIsGeneratingAi,
+    patchesSentToBackend,
+    setActiveAiPatchContainerId,
+    setActiveAiSubMaskId,
+    setError,
+  } = useAppState();
+  const setAdjustments = useSetAdjustments();
+
+  const handleGenerativeReplace = useCallback(
+    async (patchId: string, prompt: string, useFastInpaint: boolean) => {
+      if (!selectedImage?.path || isGeneratingAi) {
+        return;
+      }
+
+      const patch: AiPatch | undefined = adjustments.aiPatches.find((p: AiPatch) => p.id === patchId);
+      if (!patch) {
+        console.error('Could not find AI patch to generate for:', patchId);
+        return;
+      }
+
+      const patchDefinition = { ...patch, prompt };
+
+      setAdjustments((prev: Adjustments) => ({
+        ...prev,
+        aiPatches: prev.aiPatches.map((p: AiPatch) => (p.id === patchId ? { ...p, isLoading: true, prompt } : p)),
+      }));
+
+      setIsGeneratingAi(true);
+
+      try {
+        const newPatchDataJson: any = await invoke(Invokes.InvokeGenerativeReplaseWithMaskDef, {
+          currentAdjustments: adjustments,
+          patchDefinition: patchDefinition,
+          path: selectedImage.path,
+          useFastInpaint: useFastInpaint,
+        });
+
+        const newPatchData = JSON.parse(newPatchDataJson);
+        patchesSentToBackend.current.delete(patchId);
+        setAdjustments((prev: Adjustments) => ({
+          ...prev,
+          aiPatches: prev.aiPatches.map((p: AiPatch) =>
+            p.id === patchId
+              ? {
+                  ...p,
+                  patchData: newPatchData,
+                  isLoading: false,
+                  name: useFastInpaint ? 'Inpaint' : prompt && prompt.trim() ? prompt.trim() : p.name,
+                }
+              : p,
+          ),
+        }));
+        setActiveAiPatchContainerId(null);
+        setActiveAiSubMaskId(null);
+      } catch (err) {
+        console.error('Generative replace failed:', err);
+        setError(`AI Replace Failed: ${err}`);
+        setAdjustments((prev: Adjustments) => ({
+          ...prev,
+          aiPatches: prev.aiPatches.map((p: AiPatch) => (p.id === patchId ? { ...p, isLoading: false } : p)),
+        }));
+      } finally {
+        setIsGeneratingAi(false);
+      }
+    },
+    [
+      selectedImage?.path,
+      isGeneratingAi,
+      adjustments,
+      setAdjustments,
+      setActiveAiPatchContainerId,
+      setActiveAiSubMaskId,
+    ],
+  );
+
+  return handleGenerativeReplace;
+}
