+++ src/hooks/useGlobalEffects.ts
@@ -1,0 +1,419 @@
+import { useEffect } from 'react';
+import { useAppState } from '../context/ContextProviders';
+import { toast } from 'react-toastify';
+import { FilterCriteria, Invokes, LibraryViewMode, Panel, RawStatus, Theme } from '../components/ui/AppProperties';
+import { invoke } from '@tauri-apps/api/core';
+import { useHandlers } from './useHandlers';
+import { listen } from '@tauri-apps/api/event';
+import { getCurrentWindow } from '@tauri-apps/api/window';
+import { setTheme } from '@tauri-apps/api/app';
+import { COPYABLE_ADJUSTMENT_KEYS } from '../utils/adjustments';
+import { DEFAULT_THEME_ID, ThemeProps, THEMES } from '../utils/themes';
+import { generatePaletteFromImage } from '../utils/palette';
+import { DEBUG } from '../utils/constants';
+
+export function useGlobalEffects() {
+  const {
+    currentFolderPath,
+    preloadedDataRef,
+    imageList,
+    rootPath,
+    folderTree,
+    currentFolderPathRef,
+    isCopied,
+    setIsCopied,
+    isPasted,
+    setIsPasted,
+    error,
+    setError,
+    setAdjustments: setLiveAdjustments,
+    history,
+    activeRightPanel,
+    activeMaskContainerId,
+    activeAiPatchContainerId,
+    setIsMaskControlHovered,
+    setTransformedOriginalUrl,
+    geometricAdjustmentsKey,
+    selectedImage,
+    showOriginal,
+    transformedOriginalUrl,
+    adjustments,
+    setShowOriginal,
+    libraryViewMode,
+    setIsAIConnectorConnected,
+    setIsWindowFullScreen,
+    setAppSettings,
+    setSortCriteria,
+    setFilterCriteria,
+    setUiVisibility,
+    setLibraryViewMode,
+    setThumbnailSize,
+    setThumbnailAspectRatio,
+    setActiveTreeSection,
+    setPinnedFolderTrees,
+    isInitialMount,
+    appSettings,
+    uiVisibility,
+    thumbnailSize,
+    thumbnailAspectRatio,
+    setSupportedTypes,
+    sortCriteria,
+    filterCriteria,
+    finalPreviewUrl,
+    setAdaptivePalette,
+    theme,
+    adaptivePalette,
+    isInitialThemeMount,
+    setIsAnimatingTheme,
+    expandedFolders,
+  } = useAppState();
+
+  const { state: historyAdjustments } = history;
+
+  const { refreshImageList, generateUncroppedPreview, handleSettingsChange } = useHandlers();
+
+  useEffect(() => {
+    const handleGlobalContextMenu = (event: any) => {
+      if (!DEBUG) event.preventDefault();
+    };
+    window.addEventListener('contextmenu', handleGlobalContextMenu);
+    return () => window.removeEventListener('contextmenu', handleGlobalContextMenu);
+  }, []);
+
+  useEffect(() => {
+    if (isInitialMount.current || !appSettings || !rootPath) {
+      return;
+    }
+
+    const newFolderState = {
+      currentFolderPath,
+      expandedFolders: Array.from(expandedFolders),
+    };
+
+    if (JSON.stringify(appSettings.lastFolderState) === JSON.stringify(newFolderState)) {
+      return;
+    }
+
+    handleSettingsChange({ ...appSettings, lastFolderState: newFolderState });
+  }, [currentFolderPath, expandedFolders, rootPath, appSettings, handleSettingsChange]);
+
+  useEffect(() => {
+    if (isInitialThemeMount.current) {
+      isInitialThemeMount.current = false;
+      return;
+    }
+
+    setIsAnimatingTheme(true);
+    const timer = setTimeout(() => setIsAnimatingTheme(false), 500);
+
+    return () => clearTimeout(timer);
+  }, [theme]);
+
+  useEffect(() => {
+    const root = document.documentElement;
+    const currentThemeId = theme || DEFAULT_THEME_ID;
+
+    const baseTheme =
+      THEMES.find((t: ThemeProps) => t.id === currentThemeId) ||
+      THEMES.find((t: ThemeProps) => t.id === DEFAULT_THEME_ID);
+    if (!baseTheme) {
+      return;
+    }
+
+    let finalCssVariables: any = { ...baseTheme.cssVariables };
+    const effectThemeForWindow = baseTheme.id;
+
+    if (adaptivePalette) {
+      finalCssVariables = { ...finalCssVariables, ...adaptivePalette };
+    }
+
+    Object.entries(finalCssVariables).forEach(([key, value]) => {
+      root.style.setProperty(key, value as string);
+    });
+
+    const isLight = [Theme.Light, Theme.Snow, Theme.Arctic].includes(effectThemeForWindow);
+    invoke(Invokes.UpdateWindowEffect, { theme: isLight ? Theme.Light : Theme.Dark });
+  }, [theme, adaptivePalette]);
+
+  useEffect(() => {
+    if (appSettings?.adaptiveEditorTheme && selectedImage && finalPreviewUrl) {
+      generatePaletteFromImage(finalPreviewUrl)
+        .then(setAdaptivePalette)
+        .catch((_err) => {
+          const darkTheme = THEMES.find((t) => t.id === Theme.Dark);
+          setAdaptivePalette(darkTheme ? darkTheme.cssVariables : null);
+        });
+    } else if (!appSettings?.adaptiveEditorTheme || !selectedImage) {
+      setAdaptivePalette(null);
+    }
+  }, [appSettings?.adaptiveEditorTheme, selectedImage, finalPreviewUrl]);
+
+  useEffect(() => {
+    if (isInitialMount.current || !appSettings) {
+      return;
+    }
+    if (JSON.stringify(appSettings.filterCriteria) !== JSON.stringify(filterCriteria)) {
+      handleSettingsChange({ ...appSettings, filterCriteria });
+    }
+  }, [filterCriteria, appSettings, handleSettingsChange]);
+
+  useEffect(() => {
+    if (isInitialMount.current || !appSettings) {
+      return;
+    }
+    if (JSON.stringify(appSettings.sortCriteria) !== JSON.stringify(sortCriteria)) {
+      handleSettingsChange({ ...appSettings, sortCriteria });
+    }
+  }, [sortCriteria, appSettings, handleSettingsChange]);
+
+  useEffect(() => {
+    invoke(Invokes.GetSupportedFileTypes)
+      .then((types: any) => setSupportedTypes(types))
+      .catch((err) => console.error('Failed to load supported file types:', err));
+  }, []);
+
+  useEffect(() => {
+    if (isInitialMount.current || !appSettings) {
+      return;
+    }
+    if (appSettings.libraryViewMode !== libraryViewMode) {
+      handleSettingsChange({ ...appSettings, libraryViewMode });
+    }
+  }, [libraryViewMode, appSettings, handleSettingsChange]);
+
+  useEffect(() => {
+    if (isInitialMount.current || !appSettings) {
+      return;
+    }
+    if (appSettings.thumbnailAspectRatio !== thumbnailAspectRatio) {
+      handleSettingsChange({ ...appSettings, thumbnailAspectRatio });
+    }
+  }, [thumbnailAspectRatio, appSettings, handleSettingsChange]);
+
+  useEffect(() => {
+    if (isInitialMount.current || !appSettings) {
+      return;
+    }
+    if (appSettings.thumbnailSize !== thumbnailSize) {
+      handleSettingsChange({ ...appSettings, thumbnailSize });
+    }
+  }, [thumbnailSize, appSettings, handleSettingsChange]);
+
+  useEffect(() => {
+    if (isInitialMount.current || !appSettings) {
+      return;
+    }
+    if (JSON.stringify(appSettings.uiVisibility) !== JSON.stringify(uiVisibility)) {
+      handleSettingsChange({ ...appSettings, uiVisibility });
+    }
+  }, [uiVisibility, appSettings, handleSettingsChange]);
+
+  useEffect(() => {
+    invoke(Invokes.LoadSettings)
+      .then(async (settings: any) => {
+        if (
+          !settings.copyPasteSettings ||
+          !settings.copyPasteSettings.includedAdjustments ||
+          settings.copyPasteSettings.includedAdjustments.length === 0
+        ) {
+          settings.copyPasteSettings = {
+            mode: 'merge',
+            includedAdjustments: COPYABLE_ADJUSTMENT_KEYS,
+          };
+        }
+        setAppSettings(settings);
+        if (settings?.sortCriteria) setSortCriteria(settings.sortCriteria);
+        if (settings?.filterCriteria) {
+          setFilterCriteria((prev: FilterCriteria) => ({
+            ...prev,
+            ...settings.filterCriteria,
+            rawStatus: settings.filterCriteria.rawStatus || RawStatus.All,
+            colors: settings.filterCriteria.colors || [],
+          }));
+        }
+        if (settings?.theme) {
+          setTheme(settings.theme);
+        }
+        if (settings?.uiVisibility) {
+          setUiVisibility((prev) => ({ ...prev, ...settings.uiVisibility }));
+        }
+        if (settings?.libraryViewMode) {
+          setLibraryViewMode(settings.libraryViewMode);
+        }
+        if (settings?.thumbnailSize) {
+          setThumbnailSize(settings.thumbnailSize);
+        }
+        if (settings?.thumbnailAspectRatio) {
+          setThumbnailAspectRatio(settings.thumbnailAspectRatio);
+        }
+        if (settings?.activeTreeSection) {
+          setActiveTreeSection(settings.activeTreeSection);
+        }
+        if (settings?.pinnedFolders && settings.pinnedFolders.length > 0) {
+          try {
+            const trees = await invoke(Invokes.GetPinnedFolderTrees, { paths: settings.pinnedFolders });
+            setPinnedFolderTrees(trees);
+          } catch (err) {
+            console.error('Failed to load pinned folder trees:', err);
+          }
+        }
+
+        if (settings.lastRootPath) {
+          const root = settings.lastRootPath;
+          const currentPath = settings.lastFolderState?.currentFolderPath || root;
+
+          const command =
+            settings.libraryViewMode === LibraryViewMode.Recursive
+              ? Invokes.ListImagesRecursive
+              : Invokes.ListImagesInDir;
+
+          preloadedDataRef.current = {
+            rootPath: root,
+            currentPath: currentPath,
+            tree: invoke(Invokes.GetFolderTree, { path: root }),
+            images: invoke(command, { path: currentPath }),
+          };
+        }
+
+        invoke('frontend_ready').catch((e) => console.error('Failed to notify backend of readiness:', e));
+      })
+      .catch((err) => {
+        console.error('Failed to load settings:', err);
+        setAppSettings({ lastRootPath: null, theme: DEFAULT_THEME_ID });
+      })
+      .finally(() => {
+        isInitialMount.current = false;
+      });
+  }, []);
+
+  useEffect(() => {
+    const appWindow = getCurrentWindow();
+    const checkFullscreen = async () => {
+      setIsWindowFullScreen(await appWindow.isFullscreen());
+    };
+    checkFullscreen();
+
+    const unlistenPromise = appWindow.onResized(checkFullscreen);
+
+    return () => {
+      unlistenPromise.then((unlisten: any) => unlisten());
+    };
+  }, []);
+
+  useEffect(() => {
+    if (activeRightPanel === Panel.Crop && selectedImage?.isReady) {
+      generateUncroppedPreview(adjustments);
+    }
+  }, [adjustments, activeRightPanel, selectedImage?.isReady, generateUncroppedPreview]);
+
+  useEffect(() => {
+    const unlisten = listen('ai-connector-status-update', (event: any) => {
+      setIsAIConnectorConnected(event.payload.connected);
+    });
+    invoke(Invokes.CheckAIConnectorStatus);
+    const interval = setInterval(() => invoke(Invokes.CheckAIConnectorStatus), 10000);
+    return () => {
+      clearInterval(interval);
+      unlisten.then((f) => f());
+    };
+  }, []);
+
+  useEffect(() => {
+    if (currentFolderPath) {
+      refreshImageList();
+    }
+  }, [libraryViewMode]);
+
+  useEffect(() => {
+    if (currentFolderPath) {
+      preloadedDataRef.current = {
+        ...preloadedDataRef.current,
+        currentPath: currentFolderPath,
+        images: Promise.resolve(imageList),
+      };
+    }
+  }, [currentFolderPath, imageList]);
+
+  useEffect(() => {
+    if (rootPath && folderTree) {
+      preloadedDataRef.current = {
+        ...preloadedDataRef.current,
+        rootPath: rootPath,
+        tree: Promise.resolve(folderTree),
+      };
+    }
+  }, [rootPath, folderTree]);
+
+  useEffect(() => {
+    currentFolderPathRef.current = currentFolderPath;
+  }, [currentFolderPath]);
+
+  useEffect(() => {
+    if (!isCopied) {
+      return;
+    }
+    const timer = setTimeout(() => setIsCopied(false), 1000);
+    return () => clearTimeout(timer);
+  }, [isCopied]);
+
+  useEffect(() => {
+    if (!isPasted) {
+      return;
+    }
+    const timer = setTimeout(() => setIsPasted(false), 1000);
+    return () => clearTimeout(timer);
+  }, [isPasted]);
+
+  useEffect(() => {
+    if (error) {
+      toast.error(error);
+      setError(null);
+    }
+  }, [error]);
+
+  useEffect(() => {
+    setLiveAdjustments(historyAdjustments);
+  }, [historyAdjustments]);
+
+  useEffect(() => {
+    if (
+      (activeRightPanel !== Panel.Masks || !activeMaskContainerId) &&
+      (activeRightPanel !== Panel.Ai || !activeAiPatchContainerId)
+    ) {
+      setIsMaskControlHovered(false);
+    }
+  }, [activeRightPanel, activeMaskContainerId, activeAiPatchContainerId]);
+
+  useEffect(() => {
+    setTransformedOriginalUrl(null);
+  }, [geometricAdjustmentsKey, selectedImage?.path]);
+
+  useEffect(() => {
+    let isEffectActive = true;
+
+    const generate = async () => {
+      if (showOriginal && selectedImage?.path && !transformedOriginalUrl) {
+        try {
+          const base64Data: string = await invoke('generate_original_transformed_preview', {
+            jsAdjustments: adjustments,
+          });
+          if (isEffectActive) {
+            setTransformedOriginalUrl(base64Data);
+          }
+        } catch (e) {
+          if (isEffectActive) {
+            console.error('Failed to generate original preview:', e);
+            setError('Failed to show original image.');
+            setShowOriginal(false);
+          }
+        }
+      }
+    };
+
+    generate();
+
+    return () => {
+      isEffectActive = false;
+    };
+  }, [showOriginal, selectedImage?.path, adjustments, transformedOriginalUrl]);
+}
