+++ src/utils/constants.ts
@@ -1,1 +1,3 @@
 export const CLERK_PUBLISHABLE_KEY = 'pk_test_YnJpZWYtc2Vhc25haWwtMTIuY2xlcmsuYWNjb3VudHMuZGV2JA'; // local dev key
+
+export const DEBUG = false;
